Create Predicting car horsepower with the mtcars dataset
#Comparing linear regression models
We run two simple linear regression models, using weight and then fuel consumption to predict horsepower and compare models. From the summary of both model, it shows fuel consumption is better suited to estimate horsepower. This is because the model predicting horsepower based on fuel consumption has a higher Rsquared (0.6024) than the model that predicts based on weight(0.4339). In other words, fuel consumption predicts up to 60% of the variation in horsepower whereas weight can only explain 43%.

#Conducting multiple linear regression
We build a model using number of cylinders and miles per gallon to predict horsepower